+++ blocklets/ai-studio/api/src/libs/security.ts
@@ -2,16 +2,39 @@ import { auth } from '@blocklet/sdk/lib/middlewares';
 import { verify } from '@blocklet/sdk/lib/util/verify-sign';
 import { NextFunction, Request, Response } from 'express';
 
+import { Config } from './env';
 import logger from './logger';
 
 export const ADMIN_ROLES = ['owner', 'admin'];
 
 export const ensureAdmin = auth({ roles: ADMIN_ROLES });
 
-export const ensurePromptsEditor = auth({ roles: ['owner', 'admin', 'promptsEditor'] });
+// dynamic permission check
+export const ensurePromptsAdmin = (req: Request, res: Response, next: NextFunction) =>
+  auth({ roles: Config.serviceModePermissionMap.ensurePromptsAdminRoles })(req, res, next);
 
-export const isRefReadOnly = ({ ref, role, defaultBranch }: { ref: string; role: string; defaultBranch: string }) =>
-  ref === defaultBranch && !['admin', 'owner'].includes(role);
+// dynamic permission check
+export const ensurePromptsEditor = (req: Request, res: Response, next: NextFunction) =>
+  auth({ roles: Config.serviceModePermissionMap.ensurePromptsEditorRoles })(req, res, next);
+
+export const isRefReadOnly = ({
+  ref,
+  defaultBranch,
+  project,
+  user,
+}: {
+  ref: string;
+  defaultBranch: string;
+  project: any;
+  user: any;
+}) => {
+  if (project?.createdBy === user?.did) {
+    return false;
+  }
+  return (
+    ref === defaultBranch && ![...(Config.serviceModePermissionMap.ensurePromptsAdminRoles || [])].includes(user?.role)
+  );
+};
 
 export function ensureComponentCallOr(fallback: (req: Request, res: Response, next: NextFunction) => any) {
   return (req: Request, res: Response, next: NextFunction) => {
@@ -38,8 +61,30 @@ export function ensureComponentCallOrAdmin() {
   return ensureComponentCallOr(ensureAdmin);
 }
 
+export function ensureComponentCallOrPromptsAdmin() {
+  return ensureComponentCallOr(ensurePromptsAdmin);
+}
+
 export function ensureComponentCallOrPromptsEditor() {
   return ensureComponentCallOr(ensurePromptsEditor);
+}
+
+// dynamic permission check, not middleware
+// if component call, check sig, if not, check user role
+export function ensureComponentCallOrRolesMatch(req: Request, roles?: string[]) {
+  try {
+    const sig = req.get('x-component-sig');
+    if (sig) {
+      const verified = verify(req.body ?? {}, sig);
+      return verified;
+    }
+    if (roles) {
+      return roles.includes(req.user!.role);
+    }
+  } catch (error) {
+    // ignore error
+  }
+  return false;
 }
 
 export function ensureComponentCallOrAuth() {