+++ blocklets/ai-studio/src/contexts/session.ts
@@ -2,6 +2,8 @@ import { getDefaultBranch } from '@app/store/current-git-store';
 import { createAuthServiceSessionContext } from '@arcblock/did-connect/lib/Session';
 import { useContext } from 'react';
 
+import { Config } from '../libs/env';
+
 const { SessionProvider, SessionContext, SessionConsumer, withSession } = createAuthServiceSessionContext();
 
 export function useSessionContext(): any {
@@ -10,15 +12,27 @@ export function useSessionContext(): any {
 
 export { SessionProvider, SessionContext, SessionConsumer, withSession };
 
-export function useIsRole(...roles: string[]) {
+export function useIsRole(roles?: string[] | undefined) {
   const { session } = useSessionContext();
+  // not provided roles means no restriction
+  if (!roles) {
+    // login required
+    if (session.user?.did) {
+      return true;
+    }
+    return false;
+  }
   return roles.includes(session.user?.role);
 }
 
+export const ADMIN_ROLES = ['owner', 'admin'];
+
 export const useInitialized = () => useSessionContext().session.initialized;
 
-export const useIsAdmin = () => useIsRole('owner', 'admin');
+export const useIsAdmin = () => useIsRole(ADMIN_ROLES);
 
-export const useIsPromptEditor = () => useIsRole('owner', 'admin', 'promptsEditor');
+export const useIsPromptAdmin = () => useIsRole(Config.serviceModePermissionMap.ensurePromptsAdminRoles);
 
-export const useReadOnly = ({ ref }: { ref: string }) => !useIsAdmin() && ref === getDefaultBranch();
+export const useIsPromptEditor = () => useIsRole(Config.serviceModePermissionMap.ensurePromptsEditorRoles);
+
+export const useReadOnly = ({ ref }: { ref: string }) => !useIsPromptEditor() && ref === getDefaultBranch();