+++ blocklets/ai-studio/src/libs/env.ts
@@ -1,0 +1,11 @@
+import { ServiceMode, getServiceModePermissionMap } from '@blocklet/ai-runtime/common';
+
+export const Config = {
+  get serviceModePermissionMap() {
+    const { serviceMode } = window?.blocklet?.preferences as {
+      serviceMode: ServiceMode;
+    };
+
+    return getServiceModePermissionMap(serviceMode);
+  },
+};